fixes and improving wxwidgets 2.9 compatibility
git-svn-id: file:///home/karl/SVN2/amilab.svn/code/trunk@1808 52b0d089-85ac-4fb5-acf5-4e3172ea7aaf

+++ src/AMILab/MainFrame.h
@@ -284,8 +284,8 @@ protected:
   wxMenu *menuVisualization;
   wxMenu *menuSyntheticImages;
 
-  int usermenu_id;
-  std::map<int,std::string> usermenu_scripts; // Scripts added to the menu by the user
+  long usermenu_id;
+  std::map<long,std::string> usermenu_scripts; // Scripts added to the menu by the user
 
   CustomStatusBar* _status_bar;
